api: swap workspace/service snapshots/service test files
Created with Sturdy
