+++ api/pkg/workspaces/service/service_test.go
@@ -2,8 +2,6 @@ package service_test
 
 import (
 	"context"
-	"os"
-	"path"
 	"testing"
 
 	"getsturdy.com/api/pkg/codebases"
@@ -14,7 +12,6 @@ import (
 	db_installations "getsturdy.com/api/pkg/installations/db"
 	"getsturdy.com/api/pkg/logger"
 	module_queue "getsturdy.com/api/pkg/queue/module"
-	"getsturdy.com/api/pkg/snapshots"
 	db_snapshots "getsturdy.com/api/pkg/snapshots/db"
 	service_snapshots "getsturdy.com/api/pkg/snapshots/service"
 	db_suggestions "getsturdy.com/api/pkg/suggestions/db"
@@ -23,7 +20,6 @@ import (
 	service_view "getsturdy.com/api/pkg/view/service"
 	db_workspaces "getsturdy.com/api/pkg/workspaces/db"
 	service_workspace "getsturdy.com/api/pkg/workspaces/service"
-	"getsturdy.com/api/vcs"
 	"getsturdy.com/api/vcs/executor"
 	"getsturdy.com/api/vcs/testutil"
 
@@ -49,7 +45,7 @@ func testModule(t *testing.T) di.Module {
 		c.ImportWithForce(configuration.TestModule)
 		c.RegisterWithForce(logger.NewTest)
 
-		c.RegisterWithForce(func() *sqlx.DB { return nil })
+		c.RegisterWithForce(func() *sqlx.DB { return nil }) // make sure db is not used
 		c.Register(func() *testing.T { return t })
 		c.RegisterWithForce(testutil.TestingRepoProvider)
 	}
@@ -62,10 +58,8 @@ type testCase struct {
 	viewService      *service_view.Service
 	executorProvider executor.Provider
 
-	userID      users.ID
-	codebaseID  codebases.ID
-	workspaceID string
-	viewID      string
+	userID     users.ID
+	codebaseID codebases.ID
 }
 
 func setup(t *testing.T) *testCase {
@@ -85,68 +79,32 @@ func setup(t *testing.T) *testCase {
 	assert.NoError(t, err)
 	tc.codebaseID = cb.ID
 
-	ws, err := tc.workspaceService.Create(ctx, service_workspace.CreateWorkspaceRequest{UserID: userID, CodebaseID: cb.ID})
-	assert.NoError(t, err)
-	tc.workspaceID = ws.ID
-
-	vw, err := tc.viewService.Create(ctx, userID, ws, nil, nil)
-	assert.NoError(t, err)
-	tc.viewID = vw.ID
-
 	return tc
 }
 
-func TestSnapshot_noChanges(t *testing.T) {
+func TestCreate_no_name(t *testing.T) {
 	tc := setup(t)
 
-	snapOnce, err := tc.snapshotService.Snapshot(tc.codebaseID, tc.workspaceID, snapshots.ActionViewSync, service_snapshots.WithOnView(tc.viewID))
-	assert.NoError(t, err)
+	ctx := context.Background()
 
-	snapTwice, err := tc.snapshotService.Snapshot(tc.codebaseID, tc.workspaceID, snapshots.ActionViewSync, service_snapshots.WithOnView(tc.viewID))
+	ws, err := tc.workspaceService.Create(ctx, service_workspace.CreateWorkspaceRequest{
+		UserID:     tc.userID,
+		CodebaseID: tc.codebaseID,
+	})
 	assert.NoError(t, err)
-
-	assert.Equal(t, snapOnce.ID, snapTwice.ID, "snapshots must be identical, no changes were made")
+	assert.Equal(t, "Untitled draft", *ws.Name)
 }
 
-func TestSnapshot_noChangesDeleted(t *testing.T) {
+func TestCreate_with_name(t *testing.T) {
 	tc := setup(t)
 
-	snapOnce, err := tc.snapshotService.Snapshot(tc.codebaseID, tc.workspaceID, snapshots.ActionViewSync, service_snapshots.WithOnView(tc.viewID))
+	ctx := context.Background()
+
+	ws, err := tc.workspaceService.Create(ctx, service_workspace.CreateWorkspaceRequest{
+		UserID:     tc.userID,
+		CodebaseID: tc.codebaseID,
+		Name:       "test",
+	})
 	assert.NoError(t, err)
-
-	assert.NoError(t, tc.snapshotService.Delete(context.Background(), snapOnce))
-
-	snapTwice, err := tc.snapshotService.Snapshot(tc.codebaseID, tc.workspaceID, snapshots.ActionViewSync, service_snapshots.WithOnView(tc.viewID))
-	assert.NoError(t, err)
-
-	assert.NotEqual(t, snapOnce.ID, snapTwice.ID, "snapshots must not be identical, the first one was removed")
-}
-
-func TestSnapshot_changes(t *testing.T) {
-	tc := setup(t)
-
-	snapOnce, err := tc.snapshotService.Snapshot(tc.codebaseID, tc.workspaceID, snapshots.ActionViewSync, service_snapshots.WithOnView(tc.viewID))
-	assert.NoError(t, err)
-
-	assert.NoError(t, tc.executorProvider.New().Write(writeFile("test.txt", []byte("test"))).ExecView(tc.codebaseID, tc.viewID, "make some changes"))
-
-	snapTwice, err := tc.snapshotService.Snapshot(tc.codebaseID, tc.workspaceID, snapshots.ActionViewSync, service_snapshots.WithOnView(tc.viewID))
-	assert.NoError(t, err)
-
-	assert.NotEqual(t, snapOnce.ID, snapTwice.ID, "snapshots must not be identical, changes were made")
-}
-
-func writeFile(filename string, content []byte) func(vcs.RepoWriter) error {
-	return func(repo vcs.RepoWriter) error {
-		file, err := os.Create(path.Join(repo.Path(), filename))
-		if err != nil {
-			return err
-		}
-		defer file.Close()
-
-		if _, err := file.Write(content); err != nil {
-			return err
-		}
-		return nil
-	}
+	assert.Equal(t, "test", *ws.Name)
 }
